fix(intake): trust authoritative closing PR links
